Fix hard coded path in benchmarks
Benchmakrs have be run from the SOURCE folder to ensure that the
referenced files are found, e.g. you need to call:

    './<path-to-build-folder>/bin/benchmarks'
But in most cases you will want to modify the benchmark anyways because
the sample files are much to small anyways and you will probably want to
point to nay local file for benchmarking.

Be aware that the benchmarks are not run automatically because we
neither want nor can track any regression at this time.